@@ -1,11 +1,7 @@
-const applicationController = require('./controller.js');
+import * as applicationController from './controller.js';
 
-const setRoutes = (server) => {
+export const setRoutes = (server) => {
     server.get('/session', applicationController.createSession);
     server.post('/payment', applicationController.processPayment);
     server.get('/tokens/:merchantId', applicationController.getTokens);
-};
-
-module.exports = {
-    setRoutes
 };
