@@ -2,7 +2,7 @@
 import * as sdk from 'onlinepayments-sdk-client-js';
 
 import Logo from '../components/Logo.js';
-import StorageService from '../utilities/StorageService.js';
+import StorageService from '../services/StorageService.js';
 import Pages from '../constants/pages.js';
 import PaymentDetails from './Partials/PaymentDetails.js';
 import PaymentProductsSelection from './Partials/PaymentProductsSelection.js';
